refactor: fixtures in integration tests.
