+++ integration_tests/test_config.py
@@ -1,42 +1,14 @@
-import sys
-from contextlib import contextmanager
-from pathlib import Path
-
-import pexpect
 import pytest
 
 
-class AssertiveChild:
-    def __init__(self, child):
-        self.child = child
-
-    def expect(self, msg, timeout=3):
-        try:
-            self.child.expect(msg, timeout=timeout)
-        except pexpect.TIMEOUT as e:
-            raise AssertionError(f"Did not recieve '{msg}' within {timeout} S.") from e
-
-
-@contextmanager
-def spawn(*args, **kwargs):
-    kwargs["encoding"] = kwargs.get("encoding", "utf8")
-    child = pexpect.spawn(*args, **kwargs)
-    child.logfile = sys.stdout
-    yield AssertiveChild(child)
-    assert child.terminate(force=True), "Failed to kill process"
-
-
-config_dir = Path(__file__).parent / "config"
-
-
-def test_basic_config_loads_tidal_generates_auth_url():
+def test_basic_config_loads_tidal_generates_auth_url(spawn, config_dir):
     config = config_dir / "basic.conf"
     with spawn(f"mopidy --config {config.resolve()}") as child:
         child.expect("Connecting to TIDAL... Quality = LOSSLESS")
         child.expect("Visit link.tidal.com/.* to log in")
 
 
-def test_lazy_config_no_connect_to_tidal():
+def test_lazy_config_no_connect_to_tidal(spawn, config_dir):
     config = config_dir / "lazy.conf"
     with spawn(f"mopidy --config {config.resolve()}") as child:
         child.expect("Connecting to TIDAL... Quality = LOSSLESS")
@@ -44,7 +16,7 @@ def test_lazy_config_no_connect_to_tidal():
             child.expect("Visit link.tidal.com/.* to log in")
 
 
-def test_lazy_config_generates_auth_url_on_access():
+def test_lazy_config_generates_auth_url_on_access(spawn, config_dir):
     config = config_dir / "lazy.conf"
     with spawn(f"mopidy --config {config.resolve()}") as child:
         child.expect("Connecting to TIDAL... Quality = LOSSLESS")
